refactor(UIKIT-1275): Отрефакторен NotificationList
